requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
einops==0.8.0
fastapi==0.112.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langserve==0.2.2
langsmith==0.1.104
lightning-utilities==0.11.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
# mediapipe-silicon==0.9.2.1
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.5.5
networkx==3.3
numpy==1.26.4
openai==1.42.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
pyarrow==17.0.0
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
SQLAlchemy==2.0.32
sse-starlette==1.8.2
starlette==0.38.2
streamlit==1.37.1
sympy==1.13.2
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
torch==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
yarl==1.9.4
